# ARROW-12032: [Rust] Optimize comparison kernels