@Dandandan 2021-03-20
@jorgecarleitao FYI

I also tried related code from your `arrow2` repo, but got worse performance results there.