@yordan-pavlov 2021-03-21
@jorgecarleitao  thank you for trying out my suggestion; I am very happy to see this resulted in some tangible performance improvements. @Dandandan I love the use of godbolt.org, I should be using it more to gain further insight into what code compiles into efficient instructions.