@Dandandan 2021-03-21
I added the faster implementations using chuncks primitives to this PR and fixed the benchmarks.