@alamb 2021-03-25
> Guys this is failing to compile on my machine. It looks like the CICD failed but was merged?

 Yeah sorry @seddonm1  -- CI passed on this PR @ https://github.com/apache/arrow/pull/9759/commits/aa4a4fb25551c69fccc14e9ddffec9fbe2a35d9f but it had a logical conflict with another PR that was previously merged